refactor: rename fuzzing to diffharness
