+++ diffharness/sqlite_oracle.go
@@ -1,4 +1,4 @@
-package fuzzing
+package diffharness
 
 import (
 	"database/sql"
@@ -8,8 +8,8 @@ import (
 )
 
 // SQLiteOracle is a SQLite-backed reference implementation used by the
-// fuzzing harness. It stores every mutation with an explicit sequence number
-// so reads can be performed at past snapshots.
+// differential testing harness. It stores every mutation with an explicit
+// sequence number so reads can be performed at past snapshots.
 type SQLiteOracle struct{ db *sql.DB }
 
 // OpenSQLiteOracle opens or creates a SQLite database at the given path and
