+++ docs/dev/203/overall_plan.md
@@ -72,7 +72,7 @@ SELECT COALESCE(MAX(seq), 0) AS max_seq FROM kv;
 
 ```go
 // engine.go — your DB + the oracle adapters share this shape
-package fuzzing
+package diffharness
 
 import "context"
 
@@ -101,7 +101,7 @@ explicit sequence numbers. Implement at least the following helpers:
 
 ```go
 // sqlite_oracle.go
-package fuzzing
+package diffharness
 
 import (
 "context"
@@ -193,7 +193,7 @@ Implement these wrappers calling into your engine. If your engine exposes snapsh
 
 ```go
 // rindb_adapter.go
-package fuzzing
+package diffharness
 
 type RinDB struct {
 // your db handle(s)
@@ -247,8 +247,8 @@ return nil, nil
 ## Fuzzer & Comparator (Go)
 
 ```go
-// fuzzer.go
-package fuzzing
+// harness.go
+package diffharness
 
 import (
 "bytes"
@@ -431,7 +431,7 @@ func equal(a, b []byte) bool { /* byte-wise equality */ return bytes.Equal(a, b)
 ## Main Entrypoint
 
 ```go
-// cmd/rindb-fuzz/main.go
+// diffharness/cmd/main.go
 package main
 
 import (
@@ -441,7 +441,7 @@ import (
 "log"
 "time"
 
-"github.com/yourorg/rindb/fuzzing"
+"github.com/yourorg/rindb/diffharness"
 )
 
 func main() {
@@ -458,28 +458,28 @@ seed = int64(binary.LittleEndian.Uint64(b[:]))
 }
 log.Printf("seed=%d", seed)
 
-my, err := fuzzing.OpenRinDB("rindb-data")
+my, err := diffharness.OpenRinDB("rindb-data")
 if err != nil { log.Fatal(err) }
 defer my.Close()
 
-ref, err := fuzzing.OpenSQLiteOracle("oracle.db")
+ref, err := diffharness.OpenSQLiteOracle("oracle.db")
 if err != nil { log.Fatal(err) }
 defer ref.Close()
 
-h, err := fuzzing.NewHarness(my, ref, seed, logPath)
+h, err := diffharness.NewHarness(my, ref, seed, logPath)
 if err != nil { log.Fatal(err) }
 defer h.Close()
 
-cfg := fuzzing.Cfg{
+cfg := diffharness.Cfg{
 KeyLen:    16,
 ValLenMin: 0, ValLenMax: 1024,
 RangeMax:  1000,
-Weights: map[fuzzing.OpKind]int{
-fuzzing.OpPut:   35,
-fuzzing.OpDel:   10,
-fuzzing.OpGet:   35,
-fuzzing.OpRange: 15,
-fuzzing.OpSnap:  5,
+Weights: map[diffharness.OpKind]int{
+diffharness.OpPut:   35,
+diffharness.OpDel:   10,
+diffharness.OpGet:   35,
+diffharness.OpRange: 15,
+diffharness.OpSnap:  5,
 },
 }
 
@@ -491,7 +491,7 @@ if err := h.RunForever(cfg); err != nil { log.Fatal(err) }
 Add the missing exported method:
 
 ```go
-// in fuzzer.go
+// in harness.go
 func (h *Harness) RunForever(cfg Cfg) error { return h.runForever(cfg) }
 ```
 
@@ -545,7 +545,7 @@ Re-run the same log (same `-seed` + captured config) with **varied engine config
 
   1. **Hard crash**: `os.Exit(1)` after the op is **acknowledged** (or just before fsync for negative testing).
   2. Restart both DBs; your engine should **replay WAL**; the oracle is safe because we only append (`INSERT`).
-  3. Resume fuzzing with the **same seed**, continuing `seq` from the harness counter.
+  3. Resume differential testing with the **same seed**, continuing `seq` from the harness counter.
 
 > You can also add a **fault-injecting filesystem layer** around your I/O to simulate `EIO`, short writes, `ENOSPC`, etc.
 
@@ -583,25 +583,25 @@ require modernc.org/sqlite v1.32.0 // or latest
 **Make targets (example):**
 
 ```make
-fuzz:
-go run ./cmd/rindb-fuzz -seed=0 -log=repro.jsonl
+diffharness:
+go run ./diffharness/cmd -seed=0 -log=repro.jsonl
 
-fuzz-seed:
+diffharness-seed:
 @SEED=${SEED}; [ -z "$$SEED" ] && SEED=1; \
-go run ./cmd/rindb-fuzz -seed=$$SEED -log=repro-$$SEED.jsonl
+go run ./diffharness/cmd -seed=$$SEED -log=repro-$$SEED.jsonl
 ```
 
 ---
 
 ## Debugging Failures
 
-1. Note the error: `fuzz fail at i=123456 seq=789012 kind=OpGet: ...`
+1. Note the error: `harness fail at i=123456 seq=789012 kind=OpGet: ...`
 2. Use `repro.jsonl` (or `-seed`) to **replay** up to `i` with **verbose tracing** enabled in your engine (memtable hits, table reads, bloom checks, iterator merges).
 3. Dump **engine state summary**: manifest levels, file boundaries, tombstone counts per level, snapshot refs.
 4. If needed, implement a small **delta-debugger** to binary-search the shortest prefix that still fails.
 
 ## Testing Notes
 
-- The fuzz harness lives under `fuzzing/` and is intended solely for differential testing. All harness files reside directly in the `fuzzing` package with no subpackages.
-- Coverage and unit-test runs that generate coverage data ignore `fuzzing` to keep metrics focused on core packages.
+- The diff harness lives under `diffharness/` and is intended solely for differential testing. All harness files reside directly in the `diffharness` package with no subpackages.
+- Coverage and unit-test runs that generate coverage data ignore `diffharness` to keep metrics focused on core packages.
